refactor: modernize repository layer with typed Id and transaction management
This commit refactors the domain and persistence layers to use strongly-typed
identifiers and introduces transactional support through SessionManager.

**Domain Layer Changes:**

Repository Interfaces:
- Updated all repository method signatures to use Id instead of String:
  * findById(Id) instead of findById(String)
  * deleteById(Id) instead of deleteById(String)
  * findByUserId(Id), findByAccountId(Id), findByLoanId(Id)
- Added repository methods for atomic balance operations:
  * AccountRepository: debitBalance(Id, BigDecimal), creditBalance(Id, BigDecimal)
  * LoanRepository: decreaseRemainingAmount(Id, BigDecimal)

Entity Changes:
- Transaction: Changed amount from Long to BigDecimal, currency from String to Currency
- All entities now use Id value objects for identifiers

Use Cases:
- Updated all use cases to use typed Id for repository calls
- Refactored CreateLoan to support disbursement transactions
- RemoveTransaction now uses atomic repository operations instead of manual entity updates
- All use cases now integrate with SessionManager for transactional operations

**Test Layer Changes:**

- Updated 300+ test assertions to use Id.of() instead of raw strings
- Fixed mock verifications to expect Id parameters
- Updated CreateLoanTest with 6 new disbursement transaction test cases
- Created CreateLoanDisbursementTransactionTest with 5 test cases
- Created CreateUserTransactionTest with comprehensive transaction type coverage
- Changed amount assertions from Long to BigDecimal comparisons
- Fixed ArgumentMatchers from anyString() to any(Id.class)

**Persistence Layer Changes:**

MongoRepository Implementations:
- Implemented atomic operations using Morphia UpdateOperators:
  * debitBalance/creditBalance use inc/dec operators
  * decreaseRemainingAmount uses dec operator
- Updated all queries to use typed Id with .value() calls
- Added MONGO_ID constant ("_id") for consistency
- Updated filter predicates to use id.value() for MongoDB queries

TransactionModel:
- Changed amount field from Long to BigDecimal
- Changed currency field from String to Currency
- Updated serialization/deserialization logic

SessionManager Integration:
- MongoSessionManager now supports nested sessions
- Repository operations use pickDatastore() to select session-aware or default datastore
- Added getCurrentSession() helper with Optional return type
- Refactored executeInSession to delegate to executeNewSession when no session exists

**API Layer Changes:**

Endpoints:
- CreateTransactionEndpoint: Changed request amount type from Long to BigDecimal
- Updated all endpoint tests to use BigDecimal values
- Fixed JSON assertions to use getBigDecimal() instead of getLong()

**Key Improvements:**

1. Type Safety: Replaced stringly-typed IDs with Id value objects throughout
2. Data Precision: Use BigDecimal for monetary amounts (was Long)
3. Internationalization: Use Currency instead of String for currency codes
4. Atomicity: Repository operations now support atomic updates via MongoDB operators
5. Transaction Support: Integrated SessionManager for multi-operation transactions
6. Consistency: Standardized ID handling across all layers

**Breaking Changes:**

- All repository method signatures changed from String to Id parameters
- Transaction amount field changed from Long to BigDecimal
- Currency field changed from String to Currency type

**Files Modified:** 27
**Tests Added:** CreateLoanDisbursementTransactionTest, CreateUserTransactionTest
**Tests Updated:** All domain and integration tests (20+ files)

+++ backend/domain/src/main/java/io/dkakunsi/bitapp/transaction/dto/TransactionResult.java
@@ -1,5 +1,7 @@
 package io.dkakunsi.bitapp.transaction.dto;
 
+import java.math.BigDecimal;
+
 import lombok.Builder;
 
 @Builder
@@ -13,7 +15,7 @@ public final record TransactionResult(
     String source,
     String destination,
     String loan,
-    Long amount,
+    BigDecimal amount,
     String currency,
     String category,
     String type) {
