+++ backend/javalin/src/test/java/io/dkakunsi/bitapp/transaction/endpoint/CreateTransactionEndpointTest.java
@@ -7,6 +7,8 @@ import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 
+import java.math.BigDecimal;
+
 import org.junit.jupiter.api.AfterAll;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -56,7 +58,7 @@ class CreateTransactionEndpointTest {
         .date("2026-01-22")
         .time("10:30")
         .source("account-1")
-        .amount(50000L)
+        .amount(BigDecimal.valueOf(50000))
         .currency("IDR")
         .category("FOOD")
         .type("DEBIT")
@@ -103,7 +105,7 @@ class CreateTransactionEndpointTest {
         .date("2026-01-22")
         .time("08:00")
         .destination("account-1")
-        .amount(5000000L)
+        .amount(BigDecimal.valueOf(5000000))
         .currency("IDR")
         .category("INCOME")
         .type("CREDIT")
@@ -151,7 +153,7 @@ class CreateTransactionEndpointTest {
         .time("12:00")
         .source("account-1")
         .destination("account-2")
-        .amount(100000L)
+        .amount(BigDecimal.valueOf(100000))
         .currency("IDR")
         .category("OTHER")
         .type("TRANSFER")
@@ -197,7 +199,7 @@ class CreateTransactionEndpointTest {
         .title("Loan Payment")
         .source("account-1")
         .loan("loan-1")
-        .amount(100000L)
+        .amount(BigDecimal.valueOf(100000))
         .currency("IDR")
         .category("LOAN")
         .type("DEBIT")
